compute distance field
