+++ izzet/navmesh/distancefield.go
@@ -1,159 +1,107 @@
 package navmesh
 
-const MaxDistanceFieldValue float64 = 999999999999999
+func BuildDistanceField(chf *CompactHeightField) ([]int, int) {
+	width := chf.width
+	height := chf.height
 
-func computeDistanceTransform(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) {
-	// boundaries have a distance field of 0
-	for y := 0; y < dimensions[1]; y++ {
-		for x := 0; x < dimensions[0]; x++ {
-			voxelField[x][y][0].DistanceField = 0
-			voxelField[x][y][dimensions[2]-1].DistanceField = 0
-		}
-		for z := 0; z < dimensions[2]; z++ {
-			voxelField[0][y][z].DistanceField = 0
-			voxelField[dimensions[0]-1][y][z].DistanceField = 0
-		}
+	distances := make([]int, chf.spanCount)
+	for i := range distances {
+		distances[i] = maxDistance
 	}
 
-	for z := 1; z < dimensions[2]-1; z++ {
-		for x := 1; x < dimensions[0]-1; x++ {
-			for y := 0; y < dimensions[1]; y++ {
-				computeVoxelDistanceTransform(x, y, z, voxelField, reachField, dimensions, -1)
-			}
-		}
-	}
+	// mark boundary spans
+	for z := range height {
+		for x := range width {
+			cell := chf.cells[x+z*width]
+			spanIndex := cell.SpanIndex
+			spanCount := cell.SpanCount
 
-	for z := dimensions[2] - 2; z > 0; z-- {
-		for x := dimensions[0] - 2; x > 0; x-- {
-			for y := 0; y < dimensions[1]; y++ {
-				computeVoxelDistanceTransform(x, y, z, voxelField, reachField, dimensions, 1)
-			}
-		}
-	}
-}
-
-func computeVoxelDistanceTransform(x, y, z int, voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int, sweepDir int) {
-	xNeighbor := voxelField[x+sweepDir][y][z]
-	zNeighbor := voxelField[x][y][z+sweepDir]
-	xzNegativeNeighbor := voxelField[x-1][y][z+sweepDir]
-	xzPositiveNeighbor := voxelField[x+1][y][z+sweepDir]
-
-	xReach := reachField[x+sweepDir][y][z]
-	zReach := reachField[x][y][z+sweepDir]
-	xzNegativeReach := reachField[x-1][y][z+sweepDir]
-	xzPositiveReach := reachField[x+1][y][z+sweepDir]
-
-	var minDistanceFieldValue float64 = MaxDistanceFieldValue
-
-	if (!xNeighbor.Filled && xReach.sourceVoxel == nil) || (!zNeighbor.Filled && zReach.sourceVoxel == nil) ||
-		(!xzNegativeNeighbor.Filled && xzNegativeReach.sourceVoxel == nil) || (!xzPositiveNeighbor.Filled && xzPositiveReach.sourceVoxel == nil) {
-		minDistanceFieldValue = 0
-	} else {
-		var distanceFieldXValue float64
-		if xNeighbor.Filled {
-			distanceFieldXValue = xNeighbor.DistanceField + 1
-		} else if xReach.sourceVoxel != nil {
-			distanceFieldXValue = xReach.sourceVoxel.DistanceField + 1
-		}
-
-		var distanceFieldZValue float64
-		if zNeighbor.Filled {
-			distanceFieldZValue = zNeighbor.DistanceField + 1
-		} else if zReach.sourceVoxel != nil {
-			distanceFieldZValue = zReach.sourceVoxel.DistanceField + 1
-		}
-
-		var distanceFieldXZNegativeValue float64
-		if xzNegativeNeighbor.Filled {
-			distanceFieldXZNegativeValue = xzNegativeNeighbor.DistanceField + 1.4
-		} else if xzNegativeReach.sourceVoxel != nil {
-			distanceFieldXZNegativeValue = xzNegativeReach.sourceVoxel.DistanceField + 1.4
-		}
-
-		var distanceFieldXZPositiveValue float64
-		if xzPositiveNeighbor.Filled {
-			distanceFieldXZPositiveValue = xzPositiveNeighbor.DistanceField + 1.4
-		} else if xzPositiveReach.sourceVoxel != nil {
-			distanceFieldXZPositiveValue = xzPositiveReach.sourceVoxel.DistanceField + 1.4
-		}
-
-		if distanceFieldXValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXValue
-		}
-		if distanceFieldZValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldZValue
-		}
-		if distanceFieldXZNegativeValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXZNegativeValue
-		}
-		if distanceFieldXZPositiveValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXZPositiveValue
-		}
-	}
-
-	if minDistanceFieldValue < voxelField[x][y][z].DistanceField {
-		voxelField[x][y][z].DistanceField = minDistanceFieldValue
-	}
-}
-
-var blurWeights []float64 = []float64{
-	// 1, 2, 1,
-	// 2, 4, 2,
-	// 1, 2, 1,
-	1, 1, 1,
-	1, 1, 1,
-	1, 1, 1,
-}
-
-func blurDistanceField(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) {
-	blurredDistance := make([][][]float64, dimensions[0])
-	for i := range voxelField {
-		blurredDistance[i] = make([][]float64, dimensions[1])
-		for j := range voxelField[i] {
-			blurredDistance[i][j] = make([]float64, dimensions[2])
-		}
-	}
-
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-
-				var totalDistance float64
-				var totalWeight float64
-				for i, dir := range neighborDirs {
-					if x+dir[0] < 0 || z+dir[1] < 0 || x+dir[0] >= dimensions[0] || z+dir[1] >= dimensions[2] {
-						continue
+			for i := spanIndex; i < spanIndex+spanCount; i++ {
+				span := chf.spans[i]
+				for _, neighborIndex := range span.neighbors {
+					if neighborIndex == -1 {
+						distances[i] = 0
+						break
 					}
-
-					var neighbor *Voxel
-					reachNeighbor := &reachField[x+dir[0]][y][z+dir[1]]
-					if reachNeighbor.sourceVoxel != nil {
-						neighbor = reachNeighbor.sourceVoxel
-					} else if voxelField[x+dir[0]][y][z+dir[1]].Filled {
-						neighbor = &voxelField[x+dir[0]][y][z+dir[1]]
-					}
-
-					if neighbor == nil {
-						continue
-					}
-
-					totalWeight += blurWeights[i]
-					totalDistance += blurWeights[i] * neighbor.DistanceField
 				}
-				totalWeight += blurWeights[4]
-				totalDistance += blurWeights[4] * voxelField[x][y][z].DistanceField
-
-				totalDistance /= float64(totalWeight)
-				blurredDistance[x][y][z] = totalDistance
 			}
 		}
 	}
 
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-				voxelField[x][y][z].DistanceField = blurredDistance[x][y][z]
+	// pass 1
+	for z := range height {
+		for x := range width {
+			cell := chf.cells[x+z*width]
+			spanIndex := cell.SpanIndex
+			spanCount := cell.SpanCount
+
+			for i := spanIndex; i < spanIndex+spanCount; i++ {
+				span := chf.spans[i]
+
+				// (-1, 0)
+				if span.neighbors[dirLeft] != -1 {
+					neighborSpan := chf.spans[span.neighbors[dirLeft]]
+					distances[i] = Min(distances[i], distances[span.neighbors[dirLeft]]+2)
+
+					// (-1, -1)
+					if neighborSpan.neighbors[dirUp] != -1 {
+						distances[i] = Min(distances[i], distances[neighborSpan.neighbors[dirUp]]+3)
+					}
+				}
+
+				// (0, -1)
+				if span.neighbors[dirUp] != -1 {
+					neighborSpan := chf.spans[span.neighbors[dirUp]]
+					distances[i] = Min(distances[i], distances[span.neighbors[dirUp]]+2)
+
+					// (1, -1)
+					if neighborSpan.neighbors[dirRight] != -1 {
+						distances[i] = Min(distances[i], distances[neighborSpan.neighbors[dirRight]]+3)
+					}
+				}
 			}
 		}
 	}
+
+	// pass 2
+	for z := height - 1; z >= 0; z-- {
+		for x := height - 1; x >= 0; x-- {
+			cell := chf.cells[x+z*width]
+			spanIndex := cell.SpanIndex
+			spanCount := cell.SpanCount
+
+			for i := spanIndex; i < spanIndex+spanCount; i++ {
+				span := chf.spans[i]
+
+				// (1, 0)
+				if span.neighbors[dirRight] != -1 {
+					neighborSpan := chf.spans[span.neighbors[dirRight]]
+					distances[i] = Min(distances[i], distances[span.neighbors[dirRight]]+2)
+
+					// (1, 1)
+					if neighborSpan.neighbors[dirDown] != -1 {
+						distances[i] = Min(distances[i], distances[neighborSpan.neighbors[dirDown]]+3)
+					}
+				}
+
+				// (0, 1)
+				if span.neighbors[dirDown] != -1 {
+					neighborSpan := chf.spans[span.neighbors[dirDown]]
+					distances[i] = Min(distances[i], distances[span.neighbors[dirDown]]+2)
+
+					// (-1, 1)
+					if neighborSpan.neighbors[dirLeft] != -1 {
+						distances[i] = Min(distances[i], distances[neighborSpan.neighbors[dirLeft]]+3)
+					}
+				}
+			}
+		}
+	}
+
+	maxDist := 0
+	for i := 0; i < chf.spanCount; i++ {
+		maxDist = Max(distances[i], maxDist)
+	}
+
+	return distances, maxDist
 }
